open-logic-bit: fix unitsize in packet encoding for <32 channels
Currently, we're hard-coding the packet unitsize to 4, and sending all
data as a uint32_t. However, it seems that various output modules assume
that unitsize is the smallest byte encoding of the current number of
channels. This includes the srzip encoder, which cannot consume openlb
data, giving an error:

  Unexpected unit size, discarding logic data.

This change fixes the issue by adjusting the unitsize on generated
packets to the 1/2/3/4-byte size required for that number of channels.
We change the data buffer to a 'void *', and push 1/2/3/4-byte values as
appropriate.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>